Day 4
We're going to introduce a module here, Rebellion's ranges so make this a little easier.
writing a module to check if one range (a pair in essence) contains another or if they're connected/overlapped is a couple if statements but relying on a more feature mature and excellent module is a practice I'd encourage.